+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 * Fixed cleaved block collision sometimes bugging out
 * Reimplemented Disjunction and Vorpal to be less janky
 * Disjunction now uses the #yttr:disjunctible entity tag instead of being hardcoded
+* Control hints are no longer blurry on some GUI scales alongside QDAA